Support optional config value "discoveryInterval"
The interval in seconds at which SSDP-SEARCH packets are broadcasted to the network. The default 30s was a bit too often in a populated network. Will influence the time it takes for new devices to be found. Add e.g.
"discoveryInterval": 300
to your homebridge config.js file in the wemo platform section.